+++ src/main/java/no/finn/unleash/event/Log4JSubscriber.java
@@ -1,25 +1,40 @@
 package no.finn.unleash.event;
 
 import no.finn.unleash.UnleashException;
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.slf4j.event.Level;
+
+import static org.slf4j.event.Level.*;
 
 public class Log4JSubscriber implements UnleashSubscriber {
 
-    private static final Logger LOG = LogManager.getLogger(Log4JSubscriber.class);
+    private static final Logger LOG = LoggerFactory.getLogger(Log4JSubscriber.class);
 
-    private Level eventLevel = Level.INFO;
-    private Level errorLevel = Level.WARN;
+    private Level eventLevel = INFO;
+    private Level errorLevel = WARN;
 
     @Override
     public void on(UnleashEvent unleashEvent) {
-        LOG.log(eventLevel, unleashEvent.toString());
+        switch(eventLevel) {
+            case DEBUG: LOG.debug(unleashEvent.toString()); break;
+            case INFO: LOG.info(unleashEvent.toString()); break;
+            case WARN: LOG.warn(unleashEvent.toString()); break;
+            case ERROR: LOG.error(unleashEvent.toString()); break;
+            case TRACE: LOG.trace(unleashEvent.toString()); break;
+        }
     }
+
 
     @Override
     public void onError(UnleashException unleashException) {
-        LOG.log(errorLevel, unleashException.getMessage(), unleashException);
+        switch(errorLevel) {
+            case WARN: LOG.warn(unleashException.getMessage(), unleashException); break;
+            case ERROR: LOG.error(unleashException.getMessage(), unleashException); break;
+            case INFO: LOG.info(unleashException.getMessage(), unleashException); break;
+            case DEBUG: LOG.debug(unleashException.getMessage(), unleashException); break;
+            case TRACE: LOG.trace(unleashException.getMessage(), unleashException); break;
+        }
     }
 
     public Log4JSubscriber setEventLevel(Level eventLevel) {